connect modal, missing sorting

+++ js/book.controller.js
@@ -1,20 +1,31 @@
 'use strict'
 
 function onInit() {
-    renderBooks()
+    var lang = loadFromStorage('currLang')
+    console.log(lang);
+    if (!lang) lang = 'en-US'
+    onSetLang(lang)
+    if (lang==='he') document.querySelector('.lang-select').innerHTML =
+     '<option value="he">עברית</option> <option value="en-US">English</option>'
+     renderBooks()
 }
 
 function onSetLang(lang) {
     setLang(lang)
     renderBooks()
-
     // If lang is hebrew add RTL class to document.body
     if (lang === 'he') document.body.classList.add('rtl')
     else document.body.classList.remove('rtl')
-
-    doTrans()
 }
-
+function renderModal(bookId){
+var book = getBookById(bookId)
+console.log(book);
+document.querySelector('.modal-title').innerText = book.title
+document.querySelector('.modal-body h3').innerText = book.title
+document.querySelector('.modal-body h4').innerText = getCurrency(book.price)
+document.querySelector('img').src = `images/${book.title}.png`
+document.querySelector('p').innerText = book.desc
+}
 
 function renderBooks() {
     var books = getBooks()
@@ -25,42 +36,68 @@ function renderBooks() {
         <td class="rtl">${book.id}</td>
         <td class="rtl">${book.title}</td>
         <td class="rtl">${getCurrency(book.price)}</td>
-        <td class="read rtl"> <button type="button" class="btn btn-primary btn-sm" data-trans="read" data-bs-toggle="modal" data-bs-target="#staticBackdrop"')"> Info </button> </td>
+        <td class="read rtl"> <button type="button" onclick="renderModal('${book.id}')" class="btn btn-primary btn-sm" data-trans="read" data-bs-toggle="modal" data-bs-target="#staticBackdrop")"> Info </button> </td>
         <td class="update rtl"> <button class="btn btn-success btn-sm" data-trans="update" onclick="onUpdateBook('${book.id}')"> Update </button> </td>
         <td class="delete rtl"> <button class="btn btn-danger btn-sm" data-trans="delete" onclick="onDeleteBook('${book.id}')"> Delete </button> </td>
     </tr>`
     })
     elTable.innerHTML = strHTML
+    doTrans()
 }
 
 
 
 function onDeleteBook(bookId) {
+    var currLang = getCurrLang()
+    var confirmation = (currLang === 'he') ? confirm('לא ניתן לשחזר את הספר לאחר אישור')
+        : confirm('Book can\'t be revived after confirmation')
+
+    if (!confirmation) return
+
+    var book = getBookById(bookId)
     deleteBook(bookId)
     renderBooks()
-    flashMsg(`Book Deleted`)
+    if (currLang === 'he') flashMsg(`"${book.title}" הוסר`)
+    else flashMsg(`"${book.title}" was deleted`)
 }
 
 function onAddBook() {
-    var title = prompt('Enter title')
-    if (!title) return
-    var price = prompt('Enter price')
-    addBook(title,price)
+    var currLang = getCurrLang()
+    var title
+    var price
+    if (currLang === 'he') {
+        title = prompt('שם הספר')
+        if (!title) return
+        price = prompt('מחיר')
+    } else {
+        title = prompt('Enter title')
+        if (!title) return
+        price = prompt('Enter price')
+    }
+    addBook(title, price)
     renderBooks()
-    flashMsg(`Book Added (id: ${book.id})`)
+
+    if (currLang === 'he') flashMsg(`הספר "${title}" נוסף למאגר`)
+    else flashMsg(`"${title}" was added`)
 
 }
 
 function onUpdateBook(bookId) {
     const book = getBookById(bookId)
-    var newPrice = +prompt('Enter new price', book.price)
+    var currLang = getCurrLang()
+    var newPrice = (currLang === 'he') ? + prompt('מחיר חדש', book.price)
+        : +prompt('Enter new price', getCurrency(book.price))
+    if (newPrice === book.price) return
     if (newPrice) {
         const book = updateBook(bookId, newPrice);
         renderBooks()
-        flashMsg(`Price updated to: ${book.price}`)
     }
-}
 
+    if (currLang === 'he') flashMsg(`המחיר של "${book.title}" עודכן ל${getCurrency(book.price)}`)
+    else flashMsg(`${book.title}'s price was updated to ${getCurrency(book.price)}`)
+
+}
+getCurrency(book.price)
 function flashMsg(msg) {
     const el = document.querySelector('.user-msg')
     el.innerText = msg
